Upgraded sign in to also allow use of emailOfficial or emailPreferred
From now on, all the values of emailOfficial in the table need to be unique.  This is enforced at createPerson.  emailPreferred must also be unique in the table, and can only be set to the same value as in emailOfficial or emailPersonal, or it can be null.  emailPersonal is @unique so will be unique by SQL insert rules.

Resolves https://wevoteusa.atlassian.net/issues/WV-1077